sched_events.h: handle removal of cpu_capacity_orig
Base kernel commit 7bc263840bc3 ("sched/topology: Consolidate and clean
up access to a CPU's max compute capacity") removed
rq->cpu_capacity_orig in favor of using arch_scale_cpu_capacity() calls
in v6.7. As of v6.10 the underlying percpu variable is exported for
x86 so fix this event. In addition use SCHED_CAPACITY_SCALE in cases
where this is not available and remove compile warning when not ARM.

Note: This works on arm64 and x86. Further fixes are likely needed to
support other architectures.

Signed-off-by: Phil Auld <pauld@redhat.com>